% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{entity_time_metric_categories_df}
\alias{entity_time_metric_categories_df}
\title{Fake Data for Entity Time Metric Categories}
\format{
An object of class \code{data.frame} with 1000 rows and 21 columns.
}
\source{
fake data
}
\usage{
entity_time_metric_categories_df
}
\description{
A fake data set with 1,000 rows and 60 columns:
\describe{
\item{entity_id}{An identifier for each entity}
\item{time_column}{A time column}
\item{time_column_2}{A second time column}
\item{metric_column}{A metric column}
\item{entity_category_1}{Category for each entity}
\item{entity_category_2}{Second category for each entity}
\item{entity_category_3}{Third category for each entity}
\item{entity_outcome}{Outcome for each entity}
\item{population}{Population}
\item{spring_returned}{Boolean indicating if entity returned in the spring}
\item{second_fall_returned}{Boolean indicating if entity returned in the second fall}
\item{third_fall_returned}{Boolean indicating if entity returned in the third fall}
\item{fourth_fall_returned}{Boolean indicating if entity returned in the fourth fall}
\item{fifth_fall_returned}{Boolean indicating if entity returned in the fifth fall}
\item{sixth_fall_returned}{Boolean indicating if entity returned in the sixth fall}
\item{college}{College for each entity}
\item{department}{Department for each entity}
\item{program}{Academic program for each entity}
\item{gender}{Gender for each entity}
\item{race_ethnicity}{Race and ethnicity for each entity}
\item{cohort}{Cohort for each entity}
}
}
\keyword{datasets}
