% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sankey_diagram.R
\name{get_app_version_visibility}
\alias{get_app_version_visibility}
\title{Get version}
\usage{
get_app_version_visibility()
}
\description{
Read environment variable which decides
whether to use the public or private version
}
