% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_style.R
\name{get_title_logo}
\alias{get_title_logo}
\title{Get Title Logo}
\usage{
get_title_logo(right_aligned_title = "Shiny Dashboard")
}
\arguments{
\item{right_aligned_title}{A character string that specifies the text that will be used as the title of the app. This string will be displayed in the top right corner of the app.}
}
\value{
Returns UT logo with correct dimensions for app title.
}
\description{
Get Title Logo
}
