% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sankey_diagram.R
\name{group_choices}
\alias{group_choices}
\title{Return all group choices}
\usage{
group_choices()
}
\description{
Returns a vector of group of named column choices
Used for the Sankey drilldown options
}
