% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rate_metric_bar_chart_fct_plot_generation.R
\name{make_percent}
\alias{make_percent}
\title{Make Percent}
\usage{
make_percent(x)
}
\arguments{
\item{x}{Any real number.}
}
\value{
Input formatted as if assumed to be a percentage, with a "%" sign appended.
}
\description{
Make Percent
}
\examples{
make_percent(.5)
make_percent(1)
make_percent(90)
}
