% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_downloadable_data_table.R
\name{mod_downloadable_data_table_server}
\alias{mod_downloadable_data_table_server}
\title{downloadable_data_table Server Function}
\usage{
mod_downloadable_data_table_server(
  id,
  df = utShinyMods::entity_time_metric_categories_df,
  module_title = "Title of Module",
  module_sub_title = "Subtitle of module"
)
}
\arguments{
\item{id}{A character string giving the id of the module. This id should be unique and is used to identify the module when it is used in a Shiny app.}

\item{df}{The data frame containing the data to be displayed and available for download.}

\item{module_title}{The character string to be used as the title of the Shiny module.}

\item{module_sub_title}{The character string to be used as the subtitle of the Shiny module.}
}
\description{
This function provides the server-side logic for the `mod_downloadable_data_table` Shiny module. This function must be used in conjunction with the `mod_downloadable_data_table_ui` function in order to create a complete Shiny module.
}
