% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_downloadable_data_table.R
\name{mod_downloadable_data_table_ui}
\alias{mod_downloadable_data_table_ui}
\title{downloadable_data_table UI Function}
\usage{
mod_downloadable_data_table_ui(id)
}
\arguments{
\item{id}{A character string giving the id of the module. This id should be unique and is used to identify the module when it is used in a Shiny app.}
}
\description{
This function creates the UI portion for the `downloadable_data_table` Shiny module. This function must be used in conjunction with the `mod_downloadable_data_table_server` function in order to create a complete Shiny module.
}
