% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_help.R
\name{mod_help_server}
\alias{mod_help_server}
\title{help Server Functions}
\usage{
mod_help_server(id)
}
\description{
mod_help_ui("help_1")
}
\details{
To be copied in the server
mod_help_server("help_1")'
}
