% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_help.R
\name{mod_help_ui}
\alias{mod_help_ui}
\title{help UI Function}
\usage{
mod_help_ui(id)
}
\arguments{
\item{id}{A character string giving the id of the module. This id should be unique and is used to identify the module when it is used in a Shiny app.}
}
\description{
This function creates the UI portion for the `help` Shiny module. This function must be used in conjunction with the `mod_help_server` function in order to create a complete Shiny module.
}
