% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_over_time_line_chart.R
\name{mod_over_time_line_chart_server}
\alias{mod_over_time_line_chart_server}
\title{over_time_line_chart Server Functions}
\usage{
mod_over_time_line_chart_server(
  id,
  df = utShinyMods::entity_time_metric_categories_df,
  time_col = c(Time = "time_column"),
  metric_col = c(Metric = "metric_column"),
  metric_summarization_function = sum,
  grouping_cols = c(`Category 1` = "entity_category_1", `Category 2` =
    "entity_category_2", `Category 3` = "entity_category_3"),
  filter_cols = c(`Category 1` = "entity_category_1", `Category 2` = "entity_category_2",
    `Category 3` = "entity_category_3"),
  module_title = "Title of Module",
  module_sub_title = "Sub Title for module."
)
}
\arguments{
\item{id}{A character string giving the id of the module. This id should be unique and is used to identify the module when it is used in a Shiny app.}

\item{df}{A data frame containing the data to be plotted. The data frame should contain columns for time data, metric data, and any categorical variables to be used for grouping or filtering.}

\item{time_col}{The name of the column in `df` that contains time data.}

\item{metric_col}{The name of the column in `df` that contains metric data. This column should contain numerical data.}

\item{metric_summarization_function}{A function for summarizing the metric data. This function should take a vector of numerical data as input and return a single numerical value. The default value is `sum`, which takes the sum of the metric data.}

\item{grouping_cols}{A list of columns to group the data by. This should be a named list, with each element containing the name of a column in `df` to group by. The names of the list elements will be used as labels in the Shiny app's input controls.}

\item{filter_cols}{A list of columns to use as filters for the data. This should be a named list, with each element containing the name of a column in `df` to filter by. The names of the list elements will be used as labels in the Shiny app's input controls.}

\item{module_title}{A character string to be used as the title of the module.}

\item{module_sub_title}{A character string to be used as the sub-title of the module.}
}
\description{
This function provides the server-side logic for the `over_time_line_chart` Shiny module. This function must be used in conjunction with the `mod_over_time_line_chart_ui` function in order to create a complete Shiny module.
}
