% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_over_time_line_chart.R
\name{mod_over_time_line_chart_ui}
\alias{mod_over_time_line_chart_ui}
\title{over_time_line_chart UI Function}
\usage{
mod_over_time_line_chart_ui(id)
}
\arguments{
\item{id}{A character string giving the id of the module. This id should be unique and is used to identify the module when it is used in a Shiny app.}
}
\description{
This function creates the UI portion for the `over_time_line_chart` Shiny module. This function must be used in conjunction with the `mod_over_time_line_chart_server` function in order to create a complete Shiny module.
}
