% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rate_metric_bar_chart.R
\name{mod_rate_metric_bar_chart_server}
\alias{mod_rate_metric_bar_chart_server}
\title{rate_metric_bar_chart Server Functions
To be copied in the UI
mod_rate_metric_bar_chart_ui("rate_metric_bar_chart_1")}
\usage{
mod_rate_metric_bar_chart_server(
  id,
  df = utShinyMods::entity_time_metric_categories_df,
  time_col = c(Time = "time_column"),
  rate_metric_uniqueness_col = c(Entity = "entity_id"),
  rate_metric_criteria_col = c(Outcome = "entity_outcome"),
  rate_metric_desc = "Outcome",
  grouping_cols = c(`Category 1` = "entity_category_1", `Category 2` =
    "entity_category_2", `Category 3` = "entity_category_3"),
  module_title = "Title of Module",
  module_sub_title = "Sub Title for module."
)
}
\description{
To be copied in the server
mod_rate_metric_bar_chart_server("rate_metric_bar_chart_1")
}
