% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rate_metric_bar_chart.R
\name{mod_rate_metric_bar_chart_ui}
\alias{mod_rate_metric_bar_chart_ui}
\title{rate_metric_bar_chart UI Function}
\usage{
mod_rate_metric_bar_chart_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.

To be copied in the UI
mod_rate_metric_bar_chart_ui("rate_metric_bar_chart_1")

To be copied in the server
mod_rate_metric_bar_chart_server("rate_metric_bar_chart_1")
}
