% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sankey_diagram.R
\name{mod_sankey_diagram_server}
\alias{mod_sankey_diagram_server}
\title{sankey_diagram Server Functions}
\usage{
mod_sankey_diagram_server(
  id,
  df = utShinyMods::entity_time_metric_categories_df
)
}
\description{
To be copied in the UI
mod_sankey_diagram_ui("sankey_diagram_1")
}
\details{
To be copied in the server
mod_sankey_diagram_server("sankey_diagram_1")
}
