% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sankey_diagram.R
\name{mod_sankey_diagram_ui}
\alias{mod_sankey_diagram_ui}
\title{sankey_diagram UI Function}
\usage{
mod_sankey_diagram_ui(id)
}
\arguments{
\item{id}{A character string giving the id of the module. This id should be unique and is used to identify the module when it is used in a Shiny app.}
}
\description{
This function creates the UI portion for the `sankey_diagram` Shiny module. This function must be used in conjunction with the `mod_sankey_diagram_server` function in order to create a complete Shiny module.
}
