% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summarized_data_table.R
\name{mod_summarized_data_table_server}
\alias{mod_summarized_data_table_server}
\title{summarized_data_table Server Functions}
\usage{
mod_summarized_data_table_server(
  id,
  df = utShinyMods::entity_time_metric_categories_df,
  record_uniqueness_cols = c("entity_id", "time_column"),
  filter_col = c(Time = "time_column"),
  metric_columns = c("metric_column", "metric_column", "entity_category_1",
    "entity_category_3"),
  metric_columns_summarization_functions = c(SUM = sum, MEAN = mean, `Category 1` =
    ngram::concatenate, `Category 3` = ngram::concatenate),
  module_title = "Title of Module",
  module_sub_title = "Subtitle of module"
)
}
\description{
To be copied in the UI
mod_summarized_data_table_ui("summarized_data_table_1")
}
\details{
To be copied in the server
mod_summarized_data_table_server("summarized_data_table_1")
}
