% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summarized_data_table.R
\name{mod_summarized_data_table_ui}
\alias{mod_summarized_data_table_ui}
\title{summarized_data_table UI Function}
\usage{
mod_summarized_data_table_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
