% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sunburst_diagram.R
\name{mod_sunburst_diagram_server}
\alias{mod_sunburst_diagram_server}
\title{sunburst_diagram Server Functions}
\usage{
mod_sunburst_diagram_server(
  id,
  df = utShinyMods::entity_time_metric_categories_df,
  step_cols = c("entity_category_1", "entity_category_2", "entity_category_3",
    "entity_outcome"),
  module_title = "Title of Module",
  module_sub_title = "Sub Title for module."
)
}
\description{
To be copied in the UI
mod_sunburst_diagram_ui("sunburst_diagram_1")
}
\details{
To be copied in the server
mod_sunburst_diagram_server("sunburst_diagram_1")
}
