% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rate_metric_bar_chart_fct_plot_generation.R
\name{generate_grouped_bar_plot}
\alias{generate_grouped_bar_plot}
\title{Generate Grouped Bar Plot}
\usage{
generate_grouped_bar_plot(
  df,
  x,
  y,
  x_label,
  y_label,
  y_format = make_percent,
  grouping = "None",
  group_labeling = "",
  legend_title = "",
  legend_position = "right",
  plot_height = NULL
)
}
\arguments{
\item{plot_height}{}
}
\value{

}
\description{
Generate Grouped Bar Plot
}
