% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_over_time_line_chart_fct_plot_generation.R
\name{generate_line_chart}
\alias{generate_line_chart}
\title{Generate Line Chart}
\usage{
generate_line_chart(
  df,
  x,
  y,
  x_label,
  y_label,
  x_is_continuous = TRUE,
  x_angle = 45,
  x_format = function(x) {
     x
 },
  y_format = function(x) {
     x
 },
  grouping = 0,
  group_labeling = "",
  legend_title = "",
  legend_position = "right",
  lin_reg = FALSE
)
}
\arguments{
\item{lin_reg}{}
}
\description{
Generate Line Chart
}
