% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_over_time_line_chart_fct_plot_generation.R
\name{generate_line_chart}
\alias{generate_line_chart}
\title{Title}
\usage{
generate_line_chart(
  df,
  x,
  y,
  x_label,
  y_label,
  x_angle = 45,
  y_format = no_format,
  x_format = no_format,
  grouping = 0,
  group_labeling = "",
  legend_title = "",
  legend_position = "right",
  lin_reg = FALSE
)
}
\arguments{
\item{lin_reg}{}
}
\value{

}
\description{
Title
}
