% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_downloadable_data_table.R
\name{mod_downloadable_data_table_server}
\alias{mod_downloadable_data_table_server}
\title{downloadable_data_table Server Functions}
\usage{
mod_downloadable_data_table_server(
  id,
  df = utShinyMods::entity_time_metric_categories_df,
  module_title = "Title of Module",
  module_sub_title = "Subtitle of module"
)
}
\description{
To be copied in the UI
mod_downloadable_data_table_ui("downloadable_data_table_1")
}
\details{
To be copied in the server
mod_downloadable_data_table_server("downloadable_data_table_1")
}
