% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_downloadable_data_table.R
\name{mod_downloadable_data_table_ui}
\alias{mod_downloadable_data_table_ui}
\title{downloadable_data_table UI Function}
\usage{
mod_downloadable_data_table_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\details{
To be copied in the UI
mod_downloadable_data_table_ui("downloadable_data_table_1")

To be copied in the server
mod_downloadable_data_table_server("downloadable_data_table_1")'
}
