% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_help.R
\name{mod_help_ui}
\alias{mod_help_ui}
\title{help UI Function}
\usage{
mod_help_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\details{
To be copied in the UI
mod_help_ui("help_1")

To be copied in the server
mod_help_server("help_1")
}
