% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rate_metric_bar_chart.R
\name{mod_rate_metric_bar_chart_ui}
\alias{mod_rate_metric_bar_chart_ui}
\title{rate_metric_bar_chart UI Function}
\usage{
mod_rate_metric_bar_chart_ui(id)
}
\arguments{
\item{id}{A character string giving the id of the module. This id should be unique and is used to identify the module when it is used in a Shiny app.}
}
\description{
This function creates the UI portion for the `rate_metric_bar_chart` Shiny module. This function must be used in conjunction with the `mod_rate_metric_bar_chart_server` function in order to create a complete Shiny module.
}
