% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sankey_diagram.R
\name{mod_sankey_diagram_server}
\alias{mod_sankey_diagram_server}
\title{sankey_diagram Server Functions}
\usage{
mod_sankey_diagram_server(
  id,
  df = utShinyMods::entity_time_metric_categories_df
)
}
\arguments{
\item{id}{A character string giving the id of the module. This id should be unique and is used to identify the module when it is used in a Shiny app.}

\item{df}{A data frame containing the data that will be used to create the Sankey diagram.}
}
\description{
This function provides the server-side logic for the `sankey_diagram` Shiny module. This function must be used in conjunction with the `mod_sankey_diagram_ui` function in order to create a complete Shiny module.
}
