% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sankey_diagram.R
\name{mod_sankey_diagram_ui}
\alias{mod_sankey_diagram_ui}
\title{sankey_diagram UI Function}
\usage{
mod_sankey_diagram_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\details{
To be copied in the UI
mod_sankey_diagram_ui("sankey_diagram_1")

To be copied in the server
mod_sankey_diagram_server("sankey_diagram_1")
}
